files

+++ controllers/MailController.php
@@ -7,14 +7,18 @@ use app\models\Inbox;
 use app\models\Prepare;
 use app\models\MailForm;
 use app\models\Sent;
+use yii\helpers\Url;
 use yii\web\JsonParser;
 use Yii;
 
 class MailController extends AppController
 {
     public $layout = 'second';
-    public $obj;
-    public $ids;
+//    public $obj;
+//    public $ids;
+//    public $inboxurl;
+//    public $senturl;
+//    public $pageurl;
 
     //Функционал отправки письма и добавление новых писем в БД перенесен в виджет, см. components
 
@@ -36,26 +40,28 @@ class MailController extends AppController
     //Удаление отправленых писем
     public function actionDelete()
     {
+        $inboxurl = '/web/index.php?r=mail%2Finbox';
+        $senturl = '/web/index.php?r=mail%2Fsent';
+        $pageurl = stristr($_SERVER['HTTP_REFERER'], '/web');
+
         if (Yii::$app->request->isAjax) {
             $obj = $_POST['jsonObj'];
+//            $pageurl = $_POST['pageurl'];
         }
         $new = new Prepare(); //приведем строковое значение переменной к нужному виду
         $ids = $new->transform($obj);
-        $query = "DELETE FROM Sent WHERE id in (".$ids.")"; //создаем запрос на удаление и следом удаляем
 
-        return Yii::$app->db
-            ->createCommand($query)
-            ->queryAll();
+        if ($pageurl === $senturl) {
+            $query = "DELETE FROM Sent WHERE id in (" . $ids . ")"; //создаем запрос на удаление и следом удаляем
+
+            return Yii::$app->db
+                ->createCommand($query)
+                ->queryAll();
+        }
+
+        if ($pageurl == $inboxurl) {
+            $delmails = new Inbox();
+            $delmails->delMail($ids);
+        }
     }
-
-    public function actionDelin()
-    {
-        $host = '{imap.gmail.com:993/imap/ssl/novalidate-cert}INBOX';
-        $user = 'mailertest.dev@gmail.com';
-        $pass = 'Test123456';
-        $listmails = 1;
-        $delmails = new Inbox();
-        $delmails->delMail($host, $user, $pass, $list);
-    }
-
 }
